--- .github/workflows/php.yml
@@ -1,42 +1,0 @@
-name: PHP Composer
-
-on: [push]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    services:
-      redis:
-        image: redis
-        ports:
-        - 6379:6379
-        options: --entrypoint redis-server
-      redis2:
-        image: redis
-        ports:
-        - 6381:6379
-        options: --entrypoint redis-server
-
-    steps:
-    - uses: getong/redis-action@v1
-      with:
-        redis version: '5.0.7'
-        host port: 6380
-        container port: 6379
-        redis password: 'secret'
-    
-    - uses: actions/checkout@v1
-
-    - name: Validate composer.json and composer.lock
-      run: composer validate
-
-    - name: Install dependencies
-      run: composer install --prefer-dist --no-progress --no-suggest
-
-    # Add a test script to composer.json, for instance: "test": "vendor/bin/phpunit"
-    # Docs: https://getcomposer.org/doc/articles/scripts.md
-
-    - name: Run test suite
-      run: composer run-script test
